improve shazzrah timers
counterspell was never triggered because counterspell is only show in combatlog
if somebody was really interrupted by it. now, we check if somebody in the raid
got counterspelled and reset the timer.
minor improvements for blink timer. we cant see it in the combat log tho, so
it is most likely off after the first blink.